+++ src/main/java/com/bizcub/simpleDatapacks/AddProviders.java
@@ -2,17 +2,16 @@ package com.bizcub.simpleDatapacks;
 
 import com.bizcub.simpleDatapacks.config.Compat;
 import com.bizcub.simpleDatapacks.config.Configs;
+import net.minecraft.server.packs.PackType;
+import net.minecraft.server.packs.repository.FolderRepositorySource;
+import net.minecraft.server.packs.repository.PackSource;
+import net.minecraft.server.packs.repository.RepositorySource;
+/*? >=1.20.2*/ import net.minecraft.world.level.validation.DirectoryValidator;
 
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.server.packs.PackType;
-import net.minecraft.server.packs.repository.FolderRepositorySource;
-import net.minecraft.server.packs.repository.PackSource;
-import net.minecraft.server.packs.repository.RepositorySource;
 
 public class AddProviders {
 
@@ -30,7 +29,7 @@ public class AddProviders {
     }
 
     private static FolderRepositorySource addProvider(String path) {
-        /*? >=1.20.2*/ var validator = Minecraft.getInstance().directoryValidator();
+        /*? >=1.20.2*/ var validator = new DirectoryValidator(p -> true);
         return new FolderRepositorySource(Paths.get(path), PackType.SERVER_DATA, PackSource.DEFAULT /*? >=1.20.2 {*/, validator /*?}*/);
     }
 }