+++ src/main/java/com/bizcub/simpleDatapacks/mixin/ServerPacksSourceMixin.java
@@ -12,7 +12,6 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-
 @Mixin(ServerPacksSource.class)
 public class ServerPacksSourceMixin {