+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-* The Global Datapacks option has been removed
-* You can now specify paths to required and optional datapacks. This allows for more precise datapack customization in the game
-* To use the mod as before with the Global Datapacks option enabled, you need to move the ```datapacks``` path from optional datapacks to required datapacks
+* The mod is working on the server again